CourseAgenda refactor
