+++ src/components/organisms/CourseAgenda/CourseAgenda.tsx
@@ -1,55 +1,23 @@
-import { FC, useMemo, useContext, createContext } from "react";
+import React from "react";
 import { useTranslation } from "react-i18next";
 import styled, { withTheme } from "styled-components";
-import type { Lesson, Topic } from "@escolalms/sdk/lib/types/api";
 
-import { IconTitle, ProgressRing, Text } from "../../..";
 import { ExtendableStyledComponent } from "types/component";
-import CourseAgendaLesson from "./_components/CourseAgendaLesson";
+import { IconTitle, ProgressRing, Text } from "../../../index";
 import { ProgramIcon } from "./_components/Icons";
+import { RecursiveLessons } from "./_components/RecursiveLessons";
+import {
+  CourseAgendaContextProvider,
+  CourseAgendaContextProviderProps,
+  useCourseAgendaContext,
+} from "./context";
 
-interface TopicContext extends SharedComponentPropsHandlers {
-  lessons: Lesson[];
-  firstUnskippableTopic?: Topic;
-  positionInLessonOfFirstUnskippableTopic: number | undefined;
-  lockedLessonsIds: number[];
-  lockedTopicsIds: number[];
-}
-
-const TopicsContext = createContext<TopicContext>({
-  lessons: [],
-  lockedLessonsIds: [],
-  lockedTopicsIds: [],
-  positionInLessonOfFirstUnskippableTopic: undefined,
-  onMarkFinished: console.log,
-  onTopicClick: console.log,
-  onNextTopicClick: console.log,
-});
-
-export const useTopicsContext = () => {
-  const topicsContext = useContext(TopicsContext);
-  return topicsContext;
-};
-
-export interface SharedComponentPropsHandlers {
-  onMarkFinished: (topic: Topic) => void;
-  onTopicClick: (topic: Topic) => void;
-  onNextTopicClick: () => void;
-}
-
-export interface SharedComponentProps extends SharedComponentPropsHandlers {
+interface CourseAgendaContentProps extends ExtendableStyledComponent {
   mobile?: boolean;
-  finishedTopicIds: number[];
 }
 
-interface CourseAgendaProps
-  extends SharedComponentProps,
-    ExtendableStyledComponent {
-  lessons: Lesson[];
-  currentTopicId: number;
-  onCourseFinished: () => void;
-  unlockAllTopics?: boolean;
-}
+type CourseAgendaProps = CourseAgendaContentProps &
+  Omit<CourseAgendaContextProviderProps, "children">;
 
 const StyledSection = styled("section")`
   width: 100%;
@@ -78,142 +46,54 @@ const StyledSection = styled("section")`
       }
     }
   }
+
+  & > .lessons__list {
+    padding-left: 0;
+    list-style: none;
+  }
 `;
 
-export const CourseAgenda: FC<CourseAgendaProps> = (props) => {
-  const {
-    mobile = false,
-    lessons,
-    onMarkFinished,
-    onTopicClick,
-    onNextTopicClick,
-    className = "",
-    finishedTopicIds,
-    currentTopicId,
-    onCourseFinished,
-    unlockAllTopics,
-  } = props;
+const CourseAgendaContent: React.FC<CourseAgendaContentProps> = ({
+  mobile = false,
+  className = "",
+}) => {
   const { t } = useTranslation();
-
-  const allTopics = useMemo(
-    () => lessons.flatMap((lesson) => lesson.topics),
-    [lessons]
-  );
-
-  const firstBlockingTopic = useMemo(
-    () =>
-      allTopics
-        .filter((topic): topic is Topic => typeof topic !== "undefined")
-        .filter((topic) => !finishedTopicIds.includes(topic.id))
-        .find((topic) => !topic.can_skip),
-    [finishedTopicIds, currentTopicId]
-  );
-
-  const indexOfFirstLessonWithBlockingTopic = useMemo(
-    () =>
-      firstBlockingTopic &&
-      lessons.findIndex((lesson) => lesson.id === firstBlockingTopic.lesson_id),
-    [firstBlockingTopic]
-  );
-
-  const indexOfFirstBlockingTopic = allTopics.indexOf(firstBlockingTopic);
-
-  const positionInLessonOfFirstUnskippableTopic = useMemo(
-    () =>
-      indexOfFirstLessonWithBlockingTopic &&
-      firstBlockingTopic &&
-      lessons[indexOfFirstLessonWithBlockingTopic].topics?.indexOf(
-        firstBlockingTopic
-      ),
-    [indexOfFirstLessonWithBlockingTopic, firstBlockingTopic]
-  );
-
-  const lockedTopicsIds = useMemo(
-    () =>
-      indexOfFirstBlockingTopic !== -1
-        ? allTopics
-            .filter((_, index) => index > indexOfFirstBlockingTopic)
-            .map((topic) => (topic?.id ? topic?.id : null))
-            .filter((topicId): topicId is number => !!topicId)
-        : [],
-    [indexOfFirstBlockingTopic]
-  );
-
-  const lockedLessonsIds = useMemo(
-    () =>
-      indexOfFirstLessonWithBlockingTopic !== undefined
-        ? lessons
-            .filter((_, index) => index > indexOfFirstLessonWithBlockingTopic)
-            .map((lesson) => lesson.id)
-        : [],
-    [indexOfFirstLessonWithBlockingTopic]
-  );
-
-  // if prop 'currentTopicId' refers to locked topic, then current topic is firstBlockingTopic
-  const currentNotLockedTopicId = useMemo(
-    () =>
-      lockedTopicsIds && lockedTopicsIds.includes(currentTopicId)
-        ? firstBlockingTopic?.id
-        : currentTopicId,
-    [lockedTopicsIds, firstBlockingTopic, currentTopicId]
-  );
-
-  const percentage = useMemo(() => {
-    return Math.round((finishedTopicIds.length / allTopics.length) * 100);
-  }, [allTopics, finishedTopicIds]);
+  const { percentage, lessons } = useCourseAgendaContext();
 
   return (
     <StyledSection className={`wellms-component ${className}`}>
-      <TopicsContext.Provider
-        value={{
-          lessons: lessons,
-          firstUnskippableTopic: firstBlockingTopic,
-          positionInLessonOfFirstUnskippableTopic,
-          lockedLessonsIds,
-          lockedTopicsIds,
-          onMarkFinished,
-          onTopicClick,
-          onNextTopicClick,
-        }}
-      >
-        {!mobile && (
-          <header>
-            <IconTitle
-              level={5}
-              as="h1"
-              icon={<ProgramIcon />}
-              title={t<string>("Course.Agenda")}
-            />
-            <div>
-              <Text mode="secondary" size={"14"} noMargin>
-                {t<string>("Course.Finished")} {percentage}%
-              </Text>
-              <ProgressRing percentage={percentage} />
-            </div>
-          </header>
-        )}
-        <article>
-          {lessons.map((lesson, index) => (
-            <CourseAgendaLesson
-              unlockAllTopics={unlockAllTopics}
-              defaultOpen={lesson.topics?.some(
-                (topic) => topic.id === currentNotLockedTopicId
-              )}
-              key={lesson.id}
-              index={index}
-              {...{
-                lesson,
-                finishedTopicIds,
-                currentTopicId: currentNotLockedTopicId,
-              }}
-              onCourseFinished={onCourseFinished}
-            />
-          ))}
-        </article>
-      </TopicsContext.Provider>
+      {!mobile && (
+        <header>
+          <IconTitle
+            level={5}
+            as="h1"
+            icon={<ProgramIcon />}
+            title={t<string>("Course.Agenda")}
+          />
+          <div>
+            <Text mode="secondary" size="14" noMargin>
+              {t<string>("Course.Finished")} {percentage}%
+            </Text>
+            <ProgressRing percentage={percentage} />
+          </div>
+        </header>
+      )}
+      <ul className="lessons__list">
+        <RecursiveLessons lessons={lessons} />
+      </ul>
     </StyledSection>
   );
 };
+
+export const CourseAgenda: React.FC<CourseAgendaProps> = ({
+  className,
+  mobile,
+  ...contextProps
+}) => (
+  <CourseAgendaContextProvider {...contextProps}>
+    <CourseAgendaContent className={className} mobile={mobile} />
+  </CourseAgendaContextProvider>
+);
 
 const NewComponent = styled(CourseAgenda)<CourseAgendaProps>``;
