+++ src/styleguide/i18n.ts
@@ -19,6 +19,7 @@ export const resources = {
       "Course.Agenda": "Agenda",
       "Course.topicPreview": "Preview",
       "Course.Finished": "Finished",
+      "Course.SubLesson": "Submodule",
       "Course.Lesson": "Module",
       "Course.markAsFinished": "Mark as finished",
       "Course.nextTopic": "Switch to next topic",
@@ -269,6 +270,7 @@ export const resources = {
       "Course.Agenda": "Plan kursu",
       "Course.topicPreview": "Podgląd",
       "Course.Finished": "Ukończono",
+      "Course.SubLesson": "Podmoduł",
       "Course.Lesson": "Moduł",
       "Course.markAsFinished": "Oznacz jako ukończone",
       "Course.nextTopic": "Następne zadanie",
